Fix and enrich the `#:baseline-of` argument of `vappend`
Backward incompatibility: the default value `#f` is not recognised anymore,
and is replace with the first element of the list unless the list is empty
in which case it is 0.
I don't expect programs to break since it is currently unlikely that someone
would use `#f` explicitly. If so, it can be replaced with `0`.

`#:baseline-of` is now a mini-DSL.

Maybe it should be just called `#:baseline` now, since it recognizes
direct numbers as rows, but that would be more backward-incompatible.
Maybe I could add the `#:baseline` argument and deprecate the other one.

Fixes #1